refactor: use the mvc(s) pattern
