+++ src/main/java/com/om1cael/Main.java
@@ -1,14 +1,14 @@
 package com.om1cael;
 
-import com.om1cael.parsers.CommandParser;
-import com.om1cael.parsers.InputParser;
+import com.om1cael.view.MenuView;
+import com.om1cael.controller.InputParser;
 import com.om1cael.services.MovieVault;
 
 import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
-        CommandParser commandParser = new CommandParser(new InputParser(new Scanner(System.in)), new MovieVault());
-        commandParser.getInitialInput();
+        MenuView menuView = new MenuView(new InputParser(new Scanner(System.in)), new MovieVault());
+        menuView.getInitialInput();
     }
 }
